split HostStartupDiagnostics
simplifies the code
